refactor: move mixins and managers to separate folders
