+++ src/core/index.ts
@@ -1,32 +1,16 @@
-export { default as AccountManager } from './AccountManager.js';
-export * from './AccountManager.js';
+export { default as Session } from './Session.js';
+export * from './Session.js';
 
 export { default as Actions } from './Actions.js';
 export * from './Actions.js';
 
-export { default as Feed } from './Feed.js';
-export * from './Feed.js';
-
-export { default as FilterableFeed } from './FilterableFeed.js';
-export * from './FilterableFeed.js';
-
-export { default as InteractionManager } from './InteractionManager.js';
-export * from './InteractionManager.js';
+export { default as Player } from './Player.js';
+export * from './Player.js';
 
 export { default as OAuth } from './OAuth.js';
 export * from './OAuth.js';
 
-export { default as Player } from './Player.js';
-export * from './Player.js';
-
-export { default as PlaylistManager } from './PlaylistManager.js';
-export * from './PlaylistManager.js';
-
-export { default as Session } from './Session.js';
-export * from './Session.js';
-
-export { default as TabbedFeed } from './TabbedFeed.js';
-export * from './TabbedFeed.js';
-
 export * as Clients from './clients/index.js';
 export * as Endpoints from './endpoints/index.js';
+export * as Managers from './managers/index.js';
+export * as Mixins from './mixins/index.js';
