+++ src/core/managers/index.ts
@@ -1,0 +1,3 @@
+export { default as AccountManager } from './AccountManager.js';
+export { default as PlaylistManager } from './PlaylistManager.js';
+export { default as InteractionManager } from './InteractionManager.js';
