+++ src/core/mixins/Feed.ts
@@ -1,40 +1,40 @@
-import type { Memo, ObservedArray, SuperParsedResult, YTNode } from '../parser/helpers.js';
-import Parser, { ReloadContinuationItemsCommand } from '../parser/index.js';
-import { concatMemos, InnertubeError } from '../utils/Utils.js';
-import type Actions from './Actions.js';
+import type { Memo, ObservedArray, SuperParsedResult, YTNode } from '../../parser/helpers.js';
+import Parser, { ReloadContinuationItemsCommand } from '../../parser/index.js';
+import { concatMemos, InnertubeError } from '../../utils/Utils.js';
+import type Actions from '../Actions.js';
 
-import BackstagePost from '../parser/classes/BackstagePost.js';
-import SharedPost from '../parser/classes/SharedPost.js';
-import Channel from '../parser/classes/Channel.js';
-import CompactVideo from '../parser/classes/CompactVideo.js';
-import GridChannel from '../parser/classes/GridChannel.js';
-import GridPlaylist from '../parser/classes/GridPlaylist.js';
-import GridVideo from '../parser/classes/GridVideo.js';
-import Playlist from '../parser/classes/Playlist.js';
-import PlaylistPanelVideo from '../parser/classes/PlaylistPanelVideo.js';
-import PlaylistVideo from '../parser/classes/PlaylistVideo.js';
-import Post from '../parser/classes/Post.js';
-import ReelItem from '../parser/classes/ReelItem.js';
-import ReelShelf from '../parser/classes/ReelShelf.js';
-import RichShelf from '../parser/classes/RichShelf.js';
-import Shelf from '../parser/classes/Shelf.js';
-import Tab from '../parser/classes/Tab.js';
-import Video from '../parser/classes/Video.js';
+import BackstagePost from '../../parser/classes/BackstagePost.js';
+import SharedPost from '../../parser/classes/SharedPost.js';
+import Channel from '../../parser/classes/Channel.js';
+import CompactVideo from '../../parser/classes/CompactVideo.js';
+import GridChannel from '../../parser/classes/GridChannel.js';
+import GridPlaylist from '../../parser/classes/GridPlaylist.js';
+import GridVideo from '../../parser/classes/GridVideo.js';
+import Playlist from '../../parser/classes/Playlist.js';
+import PlaylistPanelVideo from '../../parser/classes/PlaylistPanelVideo.js';
+import PlaylistVideo from '../../parser/classes/PlaylistVideo.js';
+import Post from '../../parser/classes/Post.js';
+import ReelItem from '../../parser/classes/ReelItem.js';
+import ReelShelf from '../../parser/classes/ReelShelf.js';
+import RichShelf from '../../parser/classes/RichShelf.js';
+import Shelf from '../../parser/classes/Shelf.js';
+import Tab from '../../parser/classes/Tab.js';
+import Video from '../../parser/classes/Video.js';
 
-import AppendContinuationItemsAction from '../parser/classes/actions/AppendContinuationItemsAction.js';
-import ContinuationItem from '../parser/classes/ContinuationItem.js';
-import TwoColumnBrowseResults from '../parser/classes/TwoColumnBrowseResults.js';
-import TwoColumnSearchResults from '../parser/classes/TwoColumnSearchResults.js';
-import WatchCardCompactVideo from '../parser/classes/WatchCardCompactVideo.js';
+import AppendContinuationItemsAction from '../../parser/classes/actions/AppendContinuationItemsAction.js';
+import ContinuationItem from '../../parser/classes/ContinuationItem.js';
+import TwoColumnBrowseResults from '../../parser/classes/TwoColumnBrowseResults.js';
+import TwoColumnSearchResults from '../../parser/classes/TwoColumnSearchResults.js';
+import WatchCardCompactVideo from '../../parser/classes/WatchCardCompactVideo.js';
 
-import type MusicQueue from '../parser/classes/MusicQueue.js';
-import type RichGrid from '../parser/classes/RichGrid.js';
-import type SectionList from '../parser/classes/SectionList.js';
+import type MusicQueue from '../../parser/classes/MusicQueue.js';
+import type RichGrid from '../../parser/classes/RichGrid.js';
+import type SectionList from '../../parser/classes/SectionList.js';
 
-import type { IParsedResponse } from '../parser/types/index.js';
-import type { ApiResponse } from './Actions.js';
+import type { IParsedResponse } from '../../parser/types/index.js';
+import type { ApiResponse } from '../Actions.js';
 
-class Feed<T extends IParsedResponse = IParsedResponse> {
+export default class Feed<T extends IParsedResponse = IParsedResponse> {
   #page: T;
   #continuation?: ObservedArray<ContinuationItem>;
   #actions: Actions;
@@ -211,5 +211,3 @@ class Feed<T extends IParsedResponse = IParsedResponse> {
     return new Feed<T>(this.actions, continuation_data, true);
   }
 }
-
-export default Feed;
