+++ src/core/mixins/TabbedFeed.ts
@@ -1,13 +1,13 @@
-import Tab from '../parser/classes/Tab.js';
+import Tab from '../../parser/classes/Tab.js';
 import Feed from './Feed.js';
-import { InnertubeError } from '../utils/Utils.js';
+import { InnertubeError } from '../../utils/Utils.js';
 
-import type Actions from './Actions.js';
-import type { ObservedArray } from '../parser/helpers.js';
-import type { IParsedResponse } from '../parser/types/ParsedResponse.js';
-import type { ApiResponse } from './Actions.js';
+import type Actions from '../Actions.js';
+import type { ObservedArray } from '../../parser/helpers.js';
+import type { IParsedResponse } from '../../parser/types/ParsedResponse.js';
+import type { ApiResponse } from '../Actions.js';
 
-class TabbedFeed<T extends IParsedResponse> extends Feed<T> {
+export default class TabbedFeed<T extends IParsedResponse> extends Feed<T> {
   #tabs?: ObservedArray<Tab>;
   #actions: Actions;
 
@@ -57,5 +57,3 @@ class TabbedFeed<T extends IParsedResponse> extends Feed<T> {
     return this.page.contents_memo?.getType(Tab)?.find((tab) => tab.selected)?.title.toString();
   }
 }
-
-export default TabbedFeed;
