+++ src/core/mixins/index.ts
@@ -1,0 +1,4 @@
+export { default as Feed } from './Feed.js';
+export { default as FilterableFeed } from './FilterableFeed.js';
+export { default as TabbedFeed } from './TabbedFeed.js';
+export { default as MediaInfo } from './MediaInfo.js';
