+++ src/parser/youtube/Channel.ts
@@ -1,4 +1,4 @@
-import TabbedFeed from '../../core/TabbedFeed.js';
+import TabbedFeed from '../../core/mixins/TabbedFeed.js';
 import C4TabbedHeader from '../classes/C4TabbedHeader.js';
 import CarouselHeader from '../classes/CarouselHeader.js';
 import ChannelAboutFullMetadata from '../classes/ChannelAboutFullMetadata.js';
@@ -10,8 +10,8 @@ import ExpandableTab from '../classes/ExpandableTab.js';
 import SectionList from '../classes/SectionList.js';
 import Tab from '../classes/Tab.js';
 
-import Feed from '../../core/Feed.js';
-import FilterableFeed from '../../core/FilterableFeed.js';
+import Feed from '../../core/mixins/Feed.js';
+import FilterableFeed from '../../core/mixins/FilterableFeed.js';
 import ChipCloudChip from '../classes/ChipCloudChip.js';
 import FeedFilterChipBar from '../classes/FeedFilterChipBar.js';
 import ChannelSubMenu from '../classes/ChannelSubMenu.js';
