+++ src/parser/youtube/HomeFeed.ts
@@ -1,5 +1,5 @@
 import type Actions from '../../core/Actions.js';
-import FilterableFeed from '../../core/FilterableFeed.js';
+import FilterableFeed from '../../core/mixins/FilterableFeed.js';
 import ChipCloudChip from '../classes/ChipCloudChip.js';
 import FeedTabbedHeader from '../classes/FeedTabbedHeader.js';
 import RichGrid from '../classes/RichGrid.js';
